docker/python/requirements.txt
glob
jinja
pandas
psycopg2
typing
